4.1.0-pre3

+++ src/main/java/kr/hahaha98757/zombiesaddon/data/Window.java
@@ -4,11 +4,11 @@ package kr.hahaha98757.zombiesaddon.data;
 
 public class Window {
 	private final short[] xyz = new short[3];
-	private final int id;
+	private final byte id;
 	private boolean isActive;
 
 	public Window(int id, int x, int y, int z) {
-		this.id = id;
+		this.id = (byte) id;
 		xyz[0] = (short) x;
 		xyz[1] = (short) y;
 		xyz[2] = (short) z;
